refactor: rename minimized-glass sill buttons to .bw-pot, extract sill module
- Rename the state-named .bw-minimized-glass / .bw-minimized-detached-glass
  sill buttons to a single <button class="bw-pot" bw-plant="glass|detached-glass">,
  giving the verb pair pot (minimize) / un-pot (restore). Breaking change vs the
  react-bwin contract (.bw-minimized-glass class + getMinimizedGlassElementBySashId);
  clean rename, no alias, downstream migrates in lockstep.
- Move all sill restore logic into a new binary-window/sill.js module
  (enableSillFeatures / enableUnpotGlass / enableUnpotDetachedGlass / unpotGlass /
  getPotElementBySashId); delete detached-glass/restore.js and trim glass/action.js.
  Assemble sillModule onto BinaryWindow; enableFeatures() calls enableSillFeatures().
- CSS: merge the two sill-button rules into .bw-pot; rename keyframes/var to bw-pot.
- Document the bw- attribute-prefix convention (native els get bw-; bwin custom
  els stay unprefixed) and reframe the assemble-module TECH_DEBT fix direction.

No behavior change. Build + tests pass.
