+++ src/binary-window/detached-glass/action.minimize.js
@@ -9,16 +9,14 @@ export default {
     const sillEl = binaryWindow.sillElement;
     if (!sillEl) throw new Error(`[bwin] Sill element not found when minimizing`);
 
-    const minimizedDetachedGlassEl = createDomNode(
-      '<button class="bw-minimized-detached-glass" />'
-    );
-    sillEl.append(minimizedDetachedGlassEl);
+    const potEl = createDomNode('<button class="bw-pot" bw-plant="detached-glass" />');
+    sillEl.append(potEl);
 
     const detachedGlassEl = event.target.closest('bw-glass[detached]');
     if (!detachedGlassEl)
       throw new Error(`[bwin] Detached Glass element not found when minimizing`);
 
-    minimizedDetachedGlassEl.bwDetachedGlassElement = detachedGlassEl;
+    potEl.bwDetachedGlassElement = detachedGlassEl;
     detachedGlassEl.style.display = 'none';
   },
 };
