+++ src/binary-window/glass/action.js
@@ -1,10 +1,5 @@
-import { getMetricsFromElement } from '@/utils';
-import { getIntersectRect } from '@/rect';
-import { Position } from '@/position';
-
 export default {
   enableGlassActions() {
-    this.handleMinimizedGlassClick();
     this.observeActionButtons();
     this.dismissActionMenuOnPointerDown();
   },
@@ -19,71 +14,6 @@ export default {
       this.windowElement
         .querySelectorAll('bw-action-menu:popover-open')
         .forEach((menuEl) => menuEl.hidePopover());
-    });
-  },
-
-  restoreGlass(minimizedGlassEl) {
-    const originalRect = minimizedGlassEl.bwOriginalBoundingRect;
-
-    let biggestIntersectArea = 0;
-    let targetPaneEl = null;
-
-    this.windowElement.querySelectorAll('bw-pane').forEach((paneEl) => {
-      const paneRect = getMetricsFromElement(paneEl);
-      const intersectRect = getIntersectRect(originalRect, paneRect);
-
-      if (intersectRect) {
-        const intersectArea = intersectRect.width * intersectRect.height;
-
-        if (intersectArea > biggestIntersectArea) {
-          biggestIntersectArea = intersectArea;
-          targetPaneEl = paneEl;
-        }
-      }
-    });
-
-    if (targetPaneEl) {
-      const newPosition = minimizedGlassEl.bwOriginalPosition;
-      const targetRect = getMetricsFromElement(targetPaneEl);
-      const targetPaneSashId = targetPaneEl.getAttribute('sash-id');
-      const targetPaneSash = this.rootSash.getById(targetPaneSashId);
-
-      let newSize = 0;
-
-      if (newPosition === Position.Left || newPosition === Position.Right) {
-        newSize =
-          targetRect.width - originalRect.width < targetPaneSash.minWidth
-            ? targetRect.width / 2
-            : originalRect.width;
-      }
-      else if (newPosition === Position.Top || newPosition === Position.Bottom) {
-        newSize =
-          targetRect.height - originalRect.height < targetPaneSash.minHeight
-            ? targetRect.height / 2
-            : originalRect.height;
-      }
-      else {
-        throw new Error('[bwin] Invalid position when restoring glass');
-      }
-
-      const originalSashId = minimizedGlassEl.bwOriginalSashId;
-      const newSashPane = this.addPane(targetPaneEl.getAttribute('sash-id'), {
-        id: originalSashId,
-        position: newPosition,
-        size: newSize,
-        withGlass: false,
-      });
-      newSashPane.domNode.append(minimizedGlassEl.bwGlassElement);
-    }
-  },
-
-  handleMinimizedGlassClick() {
-    this.sillElement.addEventListener('click', (event) => {
-      if (!event.target.matches('.bw-minimized-glass')) return;
-
-      const minimizedGlassEl = event.target;
-      this.restoreGlass(minimizedGlassEl);
-      minimizedGlassEl.remove();
     });
   },
 
@@ -106,11 +36,6 @@ export default {
         el.removeAttribute('disabled');
       });
     }
-  },
-
-  getMinimizedGlassElementBySashId(sashId) {
-    const els = this.windowElement.querySelectorAll(`.bw-minimized-glass`);
-    return Array.from(els).find((el) => el.bwOriginalSashId === sashId);
   },
 
   // Re-sync action button disabled state whenever panes are added/removed,
