+++ src/binary-window/glass/action.minimize.js
@@ -9,18 +9,18 @@ export default {
     const sillEl = binaryWindow.sillElement;
     if (!sillEl) throw new Error(`[bwin] Sill element not found when minimizing`);
 
-    const minimizedGlassEl = createDomNode('<button class="bw-minimized-glass" />');
-    sillEl.append(minimizedGlassEl);
+    const potEl = createDomNode('<button class="bw-pot" bw-plant="glass" />');
+    sillEl.append(potEl);
 
     const paneEl = event.target.closest('bw-pane');
     const glassEl = event.target.closest('bw-glass');
     const paneSashId = paneEl.getAttribute('sash-id');
     const panePosition = paneEl.getAttribute('position');
 
-    minimizedGlassEl.bwGlassElement = glassEl;
-    minimizedGlassEl.bwOriginalPosition = panePosition;
-    minimizedGlassEl.bwOriginalBoundingRect = getMetricsFromElement(paneEl);
-    minimizedGlassEl.bwOriginalSashId = paneSashId;
+    potEl.bwGlassElement = glassEl;
+    potEl.bwOriginalPosition = panePosition;
+    potEl.bwOriginalBoundingRect = getMetricsFromElement(paneEl);
+    potEl.bwOriginalSashId = paneSashId;
 
     binaryWindow.removePane(paneSashId);
   },
