+++ src/css/theme.css
@@ -17,5 +17,5 @@ bw-glass[windowless][theme='dark'] {
   --bw-drop-area-bg-color: hsl(0 0% 100% / 0.1);
   --bw-detached-glass-shadow: 0 2px 6px hsl(0 0% 0% / 0.55), 0 4px 14px hsl(0 0% 0% / 0.65);
   --bw-detached-glass-shadow-active: 0 4px 10px hsl(0 0% 0% / 0.65), 0 10px 30px hsl(0 0% 0% / 0.8);
-  --bw-minimized-glass-highlight-color: hsl(0 0% 65% / 0.8);
+  --bw-pot-highlight-color: hsl(0 0% 65% / 0.8);
 }
