+++ src/css/vars.css
@@ -18,7 +18,7 @@
   --bw-glass-header-gap: 4px;
   --bw-glass-header-bg-color: hsl(0, 0%, 97%);
 
-  --bw-minimized-glass-highlight-color: hsl(0, 0%, 85%);
+  --bw-pot-highlight-color: hsl(0, 0%, 85%);
 
   --bw-detached-glass-shadow: 0 2px 6px hsl(0, 0%, 0%, 0.12), 0 4px 14px hsl(0, 0%, 0%, 0.22);
   --bw-detached-glass-shadow-active:
